ColorDisplayCreator: use VirtualStackAdapter to wrap ImagePlus
* Move code common to Color- and GrayDisplayCreator to
  AbstractDisplayCreator.
